Use p.profile instead of source.
p.profile is the chosen profile that is typed in.
It should coorespond to an aws config profile. In
the case that the profile's source profile is itself
then we have found an idp enabled role and we should
just log in to it instead of performing a second hop