Allows for form based Duo Authentication Fixes #26
In some integrations, it appears that hitting /frame/web/v1/auth with an
empty POST will not result in a 302, as expected, but returns a 200,
with a response body that contains a form with hidden input fields, sid
and certs_url. In order to continue with authentication, this form data
must be POSTed to the same /frame/web/v1/auth, which then triggers the
already handled 302 redirect flow.

This commit adds a branch to DoAuth where on StatusOK, we parse out the
values of the sid and certs_url inputs. These values are then passed
recursively to DoAuth again, so that another pass through the function
may submit these values and handle the expected 302 redirect flow. This
was found to be the order of operations needed from chrome dev tools
inspection of the iframe API calls, and it is working for our
integration setup.

It was attempted to use the form's sid value instead of trying to get it
from the 302 redirect location, but that resulted in a 403 Forbidden
from DoPrompt.

Also, GetNode has been generalized, in order to reuse the input form
parsing for sid and certs_url in addition to SAMLResponse.