Add an AWS_OKTA_PROFILE env attr
It can be useful to know which profile we're execing in, so this adds a
new env attr with that information. We don't want to use AWS_PROFILE,
since that's significant to other tools.

AWS_OKTA_PROFILE will be the name of the  profile from your config that
you're assuming.