Move session cookie handling to SAML client
This functionality does not apply for the OpenID Connect client and so can't be
abstracted by a common interface.

We still want to avoid using the session cookie in `OktaCreds.Validate`, so
we'll pass a `nil` `Keyring` to indicate that it shouldn't be checked, instead
of explicitly passing an empty session cookie.