Add OpenID Connect client for Okta
Okta APIs don't support app-level MFA. Even the [documentation for their own
AWS CLI tool][1] says:

> Note: At this time, per-app MFA is not supported in this integration. Only
organization-level MFA is supported.

This provides an alternative to the SAML client, that works better with MFA
being enabled only for a specific Okta application. With SAML it's difficult to
get the control back to the CLI program once the user has finished
authentication. With OIDC it's easier and we can then force the OS to focus the
window from which the CLI program is run, creating a reasonable smooth flow.
It's not ideal - not as good as being able to input the MFA information through
the CLI, but the best I could come up with given the APIs.

Setup instructions are included in the readme.

Fixes #63

[1]: https://support.okta.com/help/s/article/Integrating-the-Amazon-Web-Services-Command-Line-Interface-Using-Okta#Execution1