Support Yubikey configured in Duo for 2FA
This breaks a few abstractions -- and hence warrants some explanation
The Duo MFA flow in aws-okta is predicated on something like this

1. user pings okta
2. if duo, call duo API /frame/prompt which sends a push
3. this is where we have problem.. with yubikey vs push
-- instead of async waiting, yubikey can write directly to stdin and
continue, because the initial factor call wants the passcode to be
passed in
4. Then in Status call turns out the first status call returns empty
for Push but second can wait. It's opposite for Yubikey

The next step is on how to give the users a choice. I thought of few
options and settled on a flag, where the user flow is - X method
didn't work, I'll try another one. But because of the providers
nesting and calling OktaClient, it MFADevice field had to be added to
bunch of structures, where they feel very leaky. The structure doesn't
lend itself to anything more clever currently.

So the flag I introduced is --mfa-device which defaults to phone1 so
users expecting previous behavior can continue working as
is. Additionally this introduces another environment variable
AWS_OKTA_MFA_DEVICE which can be token, phone1, phone2, whatever Duo
will accept.

Alternate and a maybe architecturally cleaner way might be to ask user
which factor they want. But picking every single time when a flag
would do seems inefficient too.