fix: ensure that DoStatus is called for u2f devices (#135)
Ran into #121 while trying to get setup with Duo + U2F. After comparing the flow in the code to the browser it looks like aws-okta isn't calling the Duo status endpoint properly. That is required to exchange the txid we get for something Okta will take.

I've tested this code with both U2F and Tokens and both now work.