fix: Persist device token cookie in keystore
When using multifactor authentication Okta expects the device token
cookie `DT` in addition to the session id cookie `sid`. Without it Okta
will force you to reauthenticate the MFA session (which aws-okta
responds to by reauthenticating entirely). This is frustrating as it
means aws-okta effectively doesn't perform any caching of the Okta
signin which means the user has to reauth anytime they want to refresh /
change their aws role.

This change persists the device token the same way as the session ID.