Don't recurse when looking for assume_role_ttl
## Background

When role chaining, AWS sets a [hard upper limit of 1h on the chained assume role TTL][1].  However, the upstream non-chained role can still have a TTL of up to 12h.

## Context

Prior to this PR, if the chained role did not have a value set for `assume_role_ttl`, aws-okta would recurse up the chain looking for one. If one was set upstream with a value >1h, aws-okta would make a bad request resulting in:

```
ValidationError: The requested DurationSeconds exceeds the 1 hour session limit for roles assumed by role chaining.
	status code: 400, request id: 2e2a5e15-da98-4bd0-8457-6e95565b7fe2
```

## Change

Adds the `recursive` boolean to `GetValue` that toggles whether it will descend into `source_profile` when the given key is not found.

I've updated `updateDurationFromConfigProfile` to only look for `assume_role_ttl` in the selected profile.  I set all other instances of `GetValue` to `true` to preserve the old behavior.  However, it's worth noting that [aws-cli][2] [does not appear to recurse in this way][3].

Feedback welcome, of course.

[1]: https://github.com/awsdocs/iam-user-guide/blob/8d78057/doc_source/id_roles_terms-and-concepts.md
[2]: https://github.com/aws/aws-cli
[3]: https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-role.html