fix: make GetRoleARN support GovCloud
GovCloud support was added in #197 / b4e7839 and #204 / 9787f11.
Taking the role as an env var was added in #208 / 396d453, which added
GetRoleARN which ignored region and, thus, broke GovCloud.
Both of these were released as v0.23.0, so no version of aws-okta
supported GovCloud.

Separately, GetRoleARN was split into a lib function and a Provider
function in #218 / e13ae0f, which left behind a code duplication TODO
and an unused Provider function.

This adds Provider.GetRoleARNWithRegion.