fix: Include the profileARN for the session cache key (#237)
The profileARN can be overriden as part of the exec call.
This fixes a bug with the following commands:

- aws-okta exec profile --assume-role-arn role1 -- command
- aws-okta exec profile --assume-role-arn role2 -- command

The second command would be run using role1 since the session cache
thinks there's already a valid session. The cache lookup logic does not
take the assume role override into account when looking for a session.