feat: Persist device token cookie in keystore
Based on @Chippiewill PR #213

When using MFA, Okta expects the device token cookie `DT` in addition to
the session ID cookie `sid`. If MFA is required in the Okta app and this
cookie is not sent the session fails. This causes `aws-okta` to go
through the entire authentication flow again, eliminating any benefits
of session caching.

This change persists the device token the same way as the session ID.

I've kept the existing API identical, adding new function calls that
take a OktaCookies struct.