fix: Bug 279: Clear the DT cookie when restarting AuthN workflow (#280)
In 1.0.0 the DT (device token) cookie was added, maintaining state on
the 2nd auth factor, between client and server. However, when the Okta
session expires, the old DT was still sent. The Okta server doesn't like
when an old DT, is used with a new session and reprompts for 2FA with an
web form.

The aws-okta logic doesn't know about this 2FA web form, can't parse it,
and errors our.

We fix this bug, by clearing the DT cookie, whenever we restart the
AuthN workflow. Then the session and device token cookies are always in
sync between the client and server.